config/scripts/requirements-indexing.txt
google-auth
google-api-python-client
python-dotenv